airflow-requirements.txt
astronomer-cosmos>=1.0.2

dbt-postgres==1.8.2
fastapi
scrapy
polars